third_party/blink/web_tests/platform/generic/fast/selectors/specificity-overflow-expected.txt

PASS getComputedStyle(idVsClass).color is "rgb(0, 128, 0)"
PASS getComputedStyle(classVsElement).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE


Bug 98295: Selector specificity categories 'overflow' into higher categories

Test passes if all text on this page is green.
Test passes if all text on this page is green.
